Rename private visualization files.
