+++ optuna/multi_objective/visualization/__init__.py
@@ -1,2 +1,2 @@
 from optuna.multi_objective.visualization._pareto_front import plot_pareto_front  # NOQA
-from optuna.visualization.utils import is_available  # NOQA
+from optuna.visualization import is_available  # NOQA
