+++ optuna/multi_objective/visualization/_pareto_front.py
@@ -6,10 +6,10 @@ import optuna
 from optuna._experimental import experimental
 from optuna.multi_objective.study import MultiObjectiveStudy
 from optuna.multi_objective.trial import FrozenMultiObjectiveTrial
-from optuna.visualization.plotly_imports import _imports
+from optuna.visualization._plotly_imports import _imports
 
 if _imports.is_successful():
-    from optuna.visualization.plotly_imports import go
+    from optuna.visualization._plotly_imports import go
 
 _logger = optuna.logging.get_logger(__name__)
