+++ optuna/visualization/__init__.py
@@ -1,7 +1,7 @@
-from optuna.visualization.contour import plot_contour  # NOQA
-from optuna.visualization.intermediate_values import plot_intermediate_values  # NOQA
-from optuna.visualization.optimization_history import plot_optimization_history  # NOQA
-from optuna.visualization.parallel_coordinate import plot_parallel_coordinate  # NOQA
-from optuna.visualization.param_importances import plot_param_importances  # NOQA
-from optuna.visualization.slice import plot_slice  # NOQA
-from optuna.visualization.utils import is_available  # NOQA
+from optuna.visualization._contour import plot_contour  # NOQA
+from optuna.visualization._intermediate_values import plot_intermediate_values  # NOQA
+from optuna.visualization._optimization_history import plot_optimization_history  # NOQA
+from optuna.visualization._parallel_coordinate import plot_parallel_coordinate  # NOQA
+from optuna.visualization._param_importances import plot_param_importances  # NOQA
+from optuna.visualization._slice import plot_slice  # NOQA
+from optuna.visualization._utils import is_available  # NOQA
