+++ optuna/visualization/_parallel_coordinate.py
@@ -9,12 +9,12 @@ from optuna.logging import get_logger
 from optuna.study import Study
 from optuna.study import StudyDirection
 from optuna.trial import TrialState
-from optuna.visualization.plotly_imports import _imports
+from optuna.visualization._plotly_imports import _imports
 
 if _imports.is_successful():
-    from optuna.visualization.plotly_imports import go
+    from optuna.visualization._plotly_imports import go
 
-logger = get_logger(__name__)
+_logger = get_logger(__name__)
 
 
 def plot_parallel_coordinate(study: Study, params: Optional[List[str]] = None) -> "go.Figure":
@@ -68,7 +68,7 @@ def _get_parallel_coordinate_plot(study: Study, params: Optional[List[str]] = No
     trials = [trial for trial in study.trials if trial.state == TrialState.COMPLETE]
 
     if len(trials) == 0:
-        logger.warning("Your study does not have any completed trials.")
+        _logger.warning("Your study does not have any completed trials.")
         return go.Figure(data=[], layout=layout)
 
     all_params = {p_name for t in trials for p_name in t.params.keys()}
