+++ optuna/visualization/_utils.py
@@ -2,7 +2,7 @@ from typing import List
 
 from optuna.distributions import LogUniformDistribution
 from optuna.trial import FrozenTrial
-from optuna.visualization import plotly_imports
+from optuna.visualization import _plotly_imports
 
 __all__ = ["is_available"]
 
@@ -21,7 +21,7 @@ def is_available() -> bool:
         :obj:`True` if visualization is available, :obj:`False` otherwise.
     """
 
-    return plotly_imports._imports.is_successful()
+    return _plotly_imports._imports.is_successful()
 
 
 def _is_log_scale(trials: List[FrozenTrial], param: str) -> bool:
